requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
spacy==3.7.2
paddleocr==2.7.3.3
paddlepaddle==2.5.1
